small updates

+++ src/main/java/com/kz/pipeCutter/BBB/commands/MachinekitStop.java
@@ -1,5 +1,9 @@
 package com.kz.pipeCutter.BBB.commands;
 
+import java.util.concurrent.TimeUnit;
+
+import org.apache.commons.lang3.StringUtils;
+
 import com.jcraft.jsch.ChannelExec;
 import com.kz.pipeCutter.BBB.MyOutputStreamReader;
 import com.kz.pipeCutter.ui.Settings;
@@ -8,15 +12,29 @@ public class MachinekitStop extends SSH_Command {
 
 	@Override
 	public void runSshCmd() throws Exception {
-		String command = "ps -aux | grep 'CRAMPS.ini\\|rtapi\\|msgd\\|haltalk\\|halcmd'";
-		channelExec = (ChannelExec) session.openChannel("exec");
-		channelExec.setCommand(command);
+		
 		MyOutputStreamReader myOut = new MyOutputStreamReader();
+		channelExec = (ChannelExec) session.openChannel("exec");
 		channelExec.setOutputStream(myOut);
+
+		String avahiRestart="sudo service avahi-daemon restart";
+		
+//		channelExec.setCommand(avahiRestart);
+//		channelExec.connect(3 * 1000);
+//		while (channelExec.getExitStatus() == -1) {
+//			try {
+//				TimeUnit.SECONDS.sleep(1);
+//			} catch (Exception e) {
+//				System.out.println(e);
+//			}
+//		}
+		
+		String command = avahiRestart + "\nps -aux | grep 'CRAMPS.ini\\|rtapi\\|msgd\\|haltalk\\|halcmd'";
+		channelExec.setCommand(command);
 		channelExec.connect(3 * 1000);
 		while (channelExec.getExitStatus() == -1) {
 			try {
-				Thread.sleep(1000);
+				TimeUnit.SECONDS.sleep(1);
 			} catch (Exception e) {
 				System.out.println(e);
 			}
@@ -28,7 +46,7 @@ public class MachinekitStop extends SSH_Command {
 			if (!line.contains("grep")) {
 				String splittedLine[] = line.replaceAll("\\s+", " ").split(" ");
 				String pid = splittedLine[1];
-				if (!pid.equals(""))
+				if (!pid.equals("") && StringUtils.isNumeric(pid))
 				{
 					killCommands += "kill -9 " + pid + "\n";
 					Settings.instance.log(killCommands);
@@ -45,13 +63,16 @@ public class MachinekitStop extends SSH_Command {
 			channelExec.connect(3 * 1000);
 			while (channelExec.getExitStatus() == -1) {
 				try {
-					Thread.sleep(1000);
+					TimeUnit.SECONDS.sleep(1);
 				} catch (Exception e) {
 					System.out.println(e);
 				}
 			}
 			channelExec.disconnect();
 		}
+		
+		
+		
 
 	}
 }
